App/requirement.txt
PyPDF2
python-docx
langchain
faiss-cpu
pandas
openpyxl
torch
transformers
sentence_transformers
fastapi
uvicorn
pydantic
streamlit
requests
